app/AI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart>=0.0.6
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
xlsxwriter>=3.1.0